+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Preserve filters across task types (#265)
+
 ### Fixed
 
 -   Fixed status filter displayed value and date/duration layout (#257)